Settings: Move emergency info preference controller to correct location
