+++ src/com/android/settings/deviceinfo/EmergencyInfoPreferenceController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.android.settings.accounts;
+package com.android.settings.deviceinfo;
 
 import android.content.Context;
 import android.content.Intent;
